backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
mysql-connector-python
python-dotenv
